Add k8s version to ng status
