+++ modules/040-node-manager/hooks/discover_standby_ng.go
@@ -23,7 +23,6 @@ import (
 
 	"github.com/flant/addon-operator/pkg/module_manager/go_hook"
 	"github.com/flant/addon-operator/sdk"
-	"github.com/flant/shell-operator/pkg/kube/object_patch"
 	"github.com/flant/shell-operator/pkg/kube_events_manager/types"
 	v1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/resource"
@@ -286,15 +285,6 @@ func discoverStandbyNGHandler(input *go_hook.HookInput) error {
 
 	input.Values.Set("nodeManager.internal.standbyNodeGroups", standbyNodeGroups)
 	return nil
-}
-
-func setNodeGroupStandbyStatus(patcher *object_patch.PatchCollector, nodeGroupName string, standby *int) {
-	statusStandbyPatch := map[string]interface{}{
-		"status": map[string]interface{}{
-			"standby": standby,
-		},
-	}
-	patcher.MergePatch(statusStandbyPatch, "deckhouse.io/v1", "NodeGroup", "", nodeGroupName, object_patch.WithSubresource("/status"))
 }
 
 var NumPercentRegex = regexp.MustCompile(`^([0-9]+)%$`)
